requirements.txt
discord
mongo
bs4
requests
aiohttp
dnspython